% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.R
\name{remove_duplicated_pkgs}
\alias{remove_duplicated_pkgs}
\title{Remove duplicated packages from first element of .libPaths()}
\usage{
remove_duplicated_pkgs(
  keep = c("sf", "terra", "ragg", "svglite", "lwgeom", "PROJ")
)
}
\description{
Remove duplicated packages from first element of .libPaths()
}
